Change in BaseAuthenticator (with minor compatibility break).
A parameter "request" of type HTTPRequest has been added to the Authenticate method to permit selective access to resources.
In addition, the parameters username and password can be null in the method Authenticate, if no authorization has been received in http header. So, it is possible for the Authenticator to permit access to certain resources without any authorization information.