Serve Embedded resource : Support of space in Assembly name by replacing space to "_".
When creating a project in visual studio with space in its name, the assembly name corresponds to the project name but the Default Namespace is transform to safe project name by replacing space to "_". To access embedded resource using the assembly name, this default namespace must correspond to the assembly name. This correction prevents difficult and time-consuming resolution.